Use base class & run synchronously
